Workaround http data access in firmware
It seems the problem is related to firmware/OVMF as how it handles
"asynchronous" token. If you place the token on stack (local variable), either
the grub stack frame is corrupted, or OVMF would read out junk values from new
stack frame and misbehalving.

The problem should not happen, if we synchronized the access with the Poll() to
a flag changing its state, in order to know the token is processed and after
all grub can access it safely and dispose in sake of the task is done. It's not
likely the story, because the "disposed" token would still remain accessed by
firmware and aborted the transfer.

I also reviewed the HTTP example in UEFI spec, and did not find the difference.
Actually I am a bit sad  because I used to place the tokens on heap, and
everything working as expected.  Later on trying to use the stack after reading
the example, as it seems to be efficient and will not cause memory leak (if not
free the allocations carefully).

The measure is to place the tokens on heap or making it static (in the data or
bss section), the error is not triggered, as long as those storage will not be
reclaimed immedately for any function calls so that its more tolerate to such
fault imho.